Fix 'incorrect priviliges' bug on login
Users would receive a message along the lines
of incorrect priviliges on the first login attempt.

Sometimes, if the javascript did not load quickly
enough, users would be sent to a page to log-in, that
they did not have permission to access.  Fixed by adding
'login' to the list of allowed urls for pages.